Consolidate all Dao test to MetricsCommonDaoTests to avoid exception in PG.
