--- metrics/tests/Piipan.Metrics.Core.IntegrationTests/ParticipantSearchDaoTests.cs
@@ -1,24 +1,0 @@
-using System;
-using System.Linq;
-using System.Threading.Tasks;
-using Moq;
-using Piipan.Shared.Database;
-using Piipan.Metrics.Core.DataAccessObjects;
-using Xunit;
-using Microsoft.Extensions.Logging.Abstractions;
-using Piipan.Metrics.Core.Models;
-using Piipan.Participants.Core.Enums;
-using Piipan.Metrics.Api;
-
-namespace Piipan.Metrics.Core.IntegrationTests
-{
-    
-    public class ParticipantSearchDaoTests : DbFixture
-    {
-       
-
-     
-      
-       
-    }
-}
